+++ CHANGELOG.md
@@ -7,12 +7,13 @@
 ### Added
 
 - ✨ Filters checkboxes are using splinterlands icons when available now
-- ✨ Wins and losses in Battle History are shown in red and green now
 - ✨ Table cells are using splinterlands icons when available now
+- ✨ Wins and losses in Battle History are shown in red and green now
 
 ### Changed
 
 - 🔨 Fetching more battle data, and fetching all rule sets, user can filter by rule set in the client
+- 
   
 ## [1.2.1]